Refactor adapter, ports and services
Make adjustments to the hexagonal architecture design so it contains
a better hierarchy to support more ports and adapters. The changes
include:

* Creating new repositories to abstract interactions with other
sources besides database (ci-tron and patch archives)
* Delegating part of the fetcher code to a new a lore adapter that
is responsible for fetching the newest patches from the kernel
lore. This adapter uses the patch archive port
* This also resulted on some efforts to reduce the fetcher code
size. Unused attributes and methods were removed while the logic
for verifying for new patches was reworked - it verifies the href
of the last patch instead of storing an entire array. This method,
however, is subject to changes.