Add mocked JobManager
Implement a basic JobManager module. It provides a mocked pipeline
function that receives a patch and a database service and saves a job
marking the given patch as "COMPLETED" in the database.

It is being triggered by both the fetcher (after saving a patch) and
the TUI after selecting a saved job.

Signed-off-by: Marcelo Mendes Spessoto Junior <marcelomspessoto@gmail.com>